some fancy .obj rendering

+++ Engine/src/windows/opengl_vertexbuffer.cpp
@@ -1,7 +1,7 @@
 #include "opengl_vertexbuffer.h"
 #include "windows_defines.h"
 
-renderer::VertexBuffer::VertexBuffer(const void* data, uint8_t size)
+renderer::VertexBuffer::VertexBuffer(const void* data, size_t size)
 	: mBufferSize(size)
 {
 	glGenBuffers(1, &mVertexBufferId);
@@ -19,7 +19,7 @@ void renderer::VertexBuffer::Bind()
 	glBindBuffer(GL_ARRAY_BUFFER, mVertexBufferId);
 }
 
-std::shared_ptr<renderer::VertexBuffer> renderer::VertexBuffer::Create(const void* data, uint8_t size)
+std::shared_ptr<renderer::VertexBuffer> renderer::VertexBuffer::Create(const void* data, size_t size)
 {
 	return std::shared_ptr<renderer::VertexBuffer>(new renderer::VertexBuffer(data, size));
 }
